Added MatBlazor project reference and changed components.
MatBlazor is now installed and referenced in the Blazor.csproj. I have removed the weather and counter button from the nav bar. They still exist in the code.
I have started trying to implement a simple button, that I later think we could use in the program, but I am having issues with getting it to register a click and the css. The css is just random stand-in for now.