login, logout, register components communicate with backend api.
I cant seem to get a successful operation through the RegisterModel. The frontend and backend seem to be communicating correctly since the errors i get when trying to register match the errors from RegisterModel and RegisterResult. But something stops it from sending back a success to the frontend, thereby not registering the user.