CSS is (mostly) fixed. Changed data on site to DatasetModels.
I have changed the data being sent from ImagesGroup to being the
entire DatasetModel object. However, due to not completely
understanding matblazor tables, it seems I cannot get other data
than the ID from the datasetmodels. This is however not functionally
different from the '?' of before.
CSS is mostly fixed. I am unhappy about the last column ReviewedBy
being shorter, but cannot seem to fix the formatting. As well as
not being able to remove the header completely, which is why the box is big in the top.